feat: order is started

+++ helper/messageDeleter.ts
@@ -2,11 +2,14 @@ import { BotContext } from '#types/context'
 import { SceneFlavoredContext } from 'grammy-scenes'
 
 export const messageDeleter = async (ctx: SceneFlavoredContext<BotContext, undefined>) => {
-  for (const messageId of ctx.session.messageIds) {
-    try {
-      await ctx.api.deleteMessage(ctx.session.chatId, messageId).catch(() => {})
-    } catch (error) {
-      console.log('🚀 ~ error:', error)
-    }
+
+  if (ctx?.session?.messageIds) {
+      for (const messageId of ctx.session.messageIds) {
+        try {
+          await ctx.api.deleteMessage(ctx.session.chatId, messageId).catch(() => {})
+        } catch (error) {
+          console.log('🚀 ~ error:', error)
+        }
+      }
   }
 }
